removed cache decorators for tickets count fetch(#25)
* Add 'Manage Home' link for admin users

Introduced a new navigation link to 'Manage Home' that is visible only to users with role_id 1 (admin). This enhances navigation options for admin users in the sidebar.

* Remove caching decorators from ticket count and feed

Commented out @cache_call decorators in helpers.py for active ticket count functions and removed it from the /feed endpoint in core.py. This change ensures that updates, such as language changes, are reflected immediately without being affected by cached responses.

* Reload page after successful office deletion

Added a call to location.reload() after a successful office deletion to ensure the page reflects the latest changes.

* Comment out cache decorator for task tickets function

The @cache_call(None) decorator was commented out on get_number_of_active_tickets_task_cached, possibly to disable caching for this function. This may be for debugging or to ensure fresh data is returned.